add 3rd case of analyze
