+++ tests/resources/SNAPSHOT.analyze__main.case1.expected.soundscape.BirdNET.selection.table.txt
