+++ tests/resources/SNAPSHOT.analyze__main.case2.expected.soundscape.BirdNET.selection.table.txt
