+++ tests/resources/SNAPSHOT.analyze__main.case3.esxpected.soundscape.BirdNET.selection.table.txt
@@ -1,0 +1,30 @@
+Selection	View	Channel	Begin Time (s)	End Time (s)	Low Freq (Hz)	High Freq (Hz)	Common Name	Species Code	Confidence	Begin Path	File Offset (s)
+1	Spectrogram 1	1	0	3.0	0	15000	Schwarzkopfmeise	bkcchi	0.8141	birdnet_analyzer/example/soundscape.wav	0
+2	Spectrogram 1	1	3.0	6.0	0	15000	Schwarzkopfmeise	bkcchi	0.3083	birdnet_analyzer/example/soundscape.wav	3.0
+3	Spectrogram 1	1	6.0	9.0	0	15000	Grauhäubchenmeise	tuftit	0.1864	birdnet_analyzer/example/soundscape.wav	6.0
+4	Spectrogram 1	1	9.0	12.0	0	15000	Hausgimpel	houfin	0.6394	birdnet_analyzer/example/soundscape.wav	9.0
+5	Spectrogram 1	1	18.0	21.0	0	15000	Blauhäher	blujay	0.4353	birdnet_analyzer/example/soundscape.wav	18.0
+6	Spectrogram 1	1	21.0	24.0	0	15000	Blauhäher	blujay	0.3291	birdnet_analyzer/example/soundscape.wav	21.0
+7	Spectrogram 1	1	21.0	24.0	0	15000	Hausgimpel	houfin	0.1867	birdnet_analyzer/example/soundscape.wav	21.0
+8	Spectrogram 1	1	24.0	27.0	0	15000	Blauhäher	blujay	0.1749	birdnet_analyzer/example/soundscape.wav	24.0
+9	Spectrogram 1	1	27.0	30.0	0	15000	Winterammer	daejun	0.2187	birdnet_analyzer/example/soundscape.wav	27.0
+10	Spectrogram 1	1	33.0	36.0	0	15000	Winterammer	daejun	0.4591	birdnet_analyzer/example/soundscape.wav	33.0
+11	Spectrogram 1	1	36.0	39.0	0	15000	Winterammer	daejun	0.3537	birdnet_analyzer/example/soundscape.wav	36.0
+12	Spectrogram 1	1	39.0	42.0	0	15000	Hausgimpel	houfin	0.2533	birdnet_analyzer/example/soundscape.wav	39.0
+13	Spectrogram 1	1	42.0	45.0	0	15000	Winterammer	daejun	0.7375	birdnet_analyzer/example/soundscape.wav	42.0
+14	Spectrogram 1	1	48.0	51.0	0	15000	Blauhäher	blujay	0.1293	birdnet_analyzer/example/soundscape.wav	48.0
+15	Spectrogram 1	1	51.0	54.0	0	15000	Hausgimpel	houfin	0.2564	birdnet_analyzer/example/soundscape.wav	51.0
+16	Spectrogram 1	1	54.0	57.0	0	15000	Hausgimpel	houfin	0.6071	birdnet_analyzer/example/soundscape.wav	54.0
+17	Spectrogram 1	1	57.0	60.0	0	15000	Hausgimpel	houfin	0.1615	birdnet_analyzer/example/soundscape.wav	57.0
+18	Spectrogram 1	1	60.0	63.0	0	15000	Winterammer	daejun	0.5550	birdnet_analyzer/example/soundscape.wav	60.0
+19	Spectrogram 1	1	69.0	72.0	0	15000	Hausgimpel	houfin	0.2915	birdnet_analyzer/example/soundscape.wav	69.0
+20	Spectrogram 1	1	72.0	75.0	0	15000	Hausgimpel	houfin	0.5638	birdnet_analyzer/example/soundscape.wav	72.0
+21	Spectrogram 1	1	78.0	81.0	0	15000	Hausgimpel	houfin	0.2561	birdnet_analyzer/example/soundscape.wav	78.0
+22	Spectrogram 1	1	84.0	87.0	0	15000	Hausgimpel	houfin	0.2801	birdnet_analyzer/example/soundscape.wav	84.0
+23	Spectrogram 1	1	90.0	93.0	0	15000	Goldzeisig	amegfi	0.4190	birdnet_analyzer/example/soundscape.wav	90.0
+24	Spectrogram 1	1	93.0	96.0	0	15000	Hausgimpel	houfin	0.3168	birdnet_analyzer/example/soundscape.wav	93.0
+25	Spectrogram 1	1	96.0	99.0	0	15000	Goldzeisig	amegfi	0.4765	birdnet_analyzer/example/soundscape.wav	96.0
+26	Spectrogram 1	1	99.0	102.0	0	15000	Hausgimpel	houfin	0.1574	birdnet_analyzer/example/soundscape.wav	99.0
+27	Spectrogram 1	1	102.0	105.0	0	15000	Hausgimpel	houfin	0.4459	birdnet_analyzer/example/soundscape.wav	102.0
+28	Spectrogram 1	1	105.0	108.0	0	15000	Schwarzkopfmeise	bkcchi	0.1790	birdnet_analyzer/example/soundscape.wav	105.0
+29	Spectrogram 1	1	117.0	120.0	0	15000	Goldzeisig	amegfi	0.3924	birdnet_analyzer/example/soundscape.wav	117.0
